requirements.txt
aiogram==3.12.0
selenium
webdriver-manager
python-dotenv>=1.0.0
requests>=2.31.0
pyinstaller>=6.3.0
aiohttp>=3.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0